+++ resources/groups/resources.ps1
@@ -107,8 +107,7 @@ function Get-GoogleWSAccessToken {
         }
         $response = Invoke-RestMethod @splatParams
         $response.access_token
-    }
-    catch {
+    } catch {
         $PSCmdlet.ThrowTerminatingError($_)
     }
 }
@@ -128,7 +127,11 @@ function Invoke-GoogleWSRestMethodWithPaging {
 
         [Parameter(Mandatory)]
         [System.Collections.IDictionary]
-        $Headers
+        $Headers,
+
+        [Parameter(Mandatory)]
+        [string]
+        $CollectionName
     )
     process {
         $maxResults = 200
@@ -140,22 +143,17 @@ function Invoke-GoogleWSRestMethodWithPaging {
                 if ($Uri.Contains('?')) {
                     $urlWithOffSet = $Uri + "&maxResults=$maxResults"
                 }
-                if (-not ($null -eq $partialResult.nextPageToken)) {
-                    $urlWithOffSetAndToken = $urlWithOffSet + "&pageToken=$($partialResult.nextPageToken)"
-                } else {
-                    $urlWithOffSetAndToken = $urlWithOffSet
+                if ($partialResult.nextPageToken) {
+                    $urlWithOffset += "&pageToken=$($partialResult.nextPageToken)"
                 }
-
                 $splatParams = @{
-                    Uri         = $urlWithOffSetAndToken
-                    Headers     = $Headers
-                    Method      = $Method
-                    ContentType = $ContentType
+                    Uri     = $urlWithOffset
+                    Headers = $Headers
+                    Method  = $Method
                 }
                 $partialResult = Invoke-RestMethod @splatParams -Verbose:$false
-
-                if ($partialResult.groups.Count -gt 1) {
-                    $returnList.AddRange($partialResult.groups)
+                if ($partialResult.$CollectionName.Count -gt 0) {
+                    $returnList.AddRange($partialResult.$CollectionName)
                 }
 
             } until ($null -eq $partialResult.nextPageToken)
@@ -197,7 +195,7 @@ try {
         Method  = 'GET'
         Headers = $headers
     }
-    $googleGroups = Invoke-GoogleWSRestMethodWithPaging @splatGetGroups
+    $googleGroups = Invoke-GoogleWSRestMethodWithPaging @splatGetGroups -CollectionName 'groups'
     $googleGroupsGrouped = $googleGroups | Group-Object -Property email -AsHashTable -AsString
     if ($null -eq $googleGroupsGrouped) {
         $googleGroupsGrouped = @{}
@@ -219,9 +217,9 @@ try {
                         Headers     = $headers
                         ContentType = 'application/json'
                         Body        = ([ordered]@{
-                            name  = $resource
-                            email = $emailFormatted
-                        }
+                                name  = $resource
+                                email = $emailFormatted
+                            }
                         ) | ConvertTo-Json
                     }
                     $null = Invoke-RestMethod @splatSetGroup