+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [1.1.0] - 28-05-2025
+
+### Added
+- Added Account Import script
+- Added Group Permissions Import script
+- Added Licenses Permissions Import script
+- Added `Collection` parameter in function `Invoke-GoogleWSRestMethodWithPaging`
+
+### Changed
+- Bug Fix paging counter in `Invoke-GoogleWSRestMethodWithPaging`
+- Bug Fix account object create/update script
+
 ## [1.0.0] - 21-01-2025
 
 This is the first official release of _HelloID-Conn-Prov-Target-GoogleWorkSpace_. This release is based on template version _2.0.1_.